Move child IDs into identity
