+++ crates/shelterwood/src/policy.rs
@@ -2,6 +2,8 @@
 
 use std::{fmt, num::NonZeroUsize, time::Duration};
 
+use crate::identity::ChildId;
+
 /// Whether a scope has fixed ordered membership or runtime-dynamic membership.
 #[derive(Clone, Copy, Debug, Eq, PartialEq)]
 pub(crate) enum ScopeFlavor {
@@ -15,50 +17,6 @@ pub const DEFAULT_MAILBOX_CAPACITY: usize = 64;
 pub const DEFAULT_SHUTDOWN_GRACE: Duration = Duration::from_secs(5);
 /// The default gated-readiness deadline.
 pub const DEFAULT_READINESS_DEADLINE: Duration = Duration::from_secs(30);
-
-/// A child identifier within one scope.
-#[derive(Clone, Debug, Eq, Hash, Ord, PartialEq, PartialOrd)]
-pub struct ChildId(String);
-
-impl ChildId {
-    pub(crate) fn validate(value: impl Into<String>) -> Result<Self, IdError> {
-        let value = value.into();
-        if value.is_empty() {
-            Err(IdError::Empty)
-        } else {
-            Ok(Self(value))
-        }
-    }
-
-    /// Returns the identifier as text.
-    #[must_use]
-    pub fn as_str(&self) -> &str {
-        &self.0
-    }
-}
-
-impl fmt::Display for ChildId {
-    fn fmt(&self, formatter: &mut fmt::Formatter<'_>) -> fmt::Result {
-        self.0.fmt(formatter)
-    }
-}
-
-impl From<&str> for ChildId {
-    fn from(value: &str) -> Self {
-        Self(value.to_owned())
-    }
-}
-
-impl From<String> for ChildId {
-    fn from(value: String) -> Self {
-        Self(value)
-    }
-}
-
-#[derive(Clone, Copy, Debug, Eq, PartialEq)]
-pub(crate) enum IdError {
-    Empty,
-}
 
 /// The condition under which an exited child is restarted.
 #[derive(Clone, Copy, Debug, Eq, Hash, PartialEq)]
